+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added assert to NetCDF4_get_var.H to print variable name if data retrieval fails
 
-- Added MAPL_GetGlobalHorzIJIndex subroutine
+- Added subroutine to calculate IJ indexes of given lats lons for cubed-sphere grid
 
 ### Changed